requirements.txt
Django==4.2.11
djangorestframework==3.15.1
drf-yasg==1.21.7

pytest==8.2.0
pytest-django==4.8.0
pytest-xdist==3.6.1
